refactor: 예약 목록 조회 API를 admin 아래로 이동
